+++ x-pack/plugins/maps/public/classes/sources/ems_file_source/ems_file_source.tsx
@@ -10,7 +10,7 @@ import { Feature } from 'geojson';
 import { Adapters } from 'src/plugins/inspector/public';
 import { FileLayer } from '@elastic/ems-client';
 import { Attribution, ImmutableSourceProperty, SourceEditorArgs } from '../source';
-import { AbstractVectorSource, GeoJsonWithMeta, IVectorSource } from '../vector_source';
+import { AbstractVectorSource, GeoJsonWithMeta, IVectorSource, getGeoJson } from '../vector_source';
 import { SOURCE_TYPES, FIELD_ORIGIN, VECTOR_SHAPE_TYPE } from '../../../../common/constants';
 import { getEmsFileLayers } from '../../../meta';
 import { getDataSourceLabel } from '../../../../common/i18n_getters';
@@ -100,8 +100,7 @@ export class EMSFileSource extends AbstractVectorSource implements IEmsFileSourc
 
   async getGeoJsonWithMeta(): Promise<GeoJsonWithMeta> {
     const emsFileLayer = await this.getEMSFileLayer();
-    // @ts-ignore
-    const featureCollection = await AbstractVectorSource.getGeoJson({
+    const featureCollection = await getGeoJson({
       format: emsFileLayer.getDefaultFormatType(),
       featureCollectionPath: 'data',
       fetchUrl: emsFileLayer.getDefaultFormatUrl(),
